Extract phase1_viewer capture layer to viewer_capture.py; delete the renderer
Per session feedback, the existing top-down match_viewer.py is the
canonical dev viewer; the Phase 1 anatomical viewer never read as
well. Rather than ship the dead viewer to main, extract its
genuinely-reusable bits and drop the rest:

KEPT (now in src/viewer_capture.py — pure data, no pygame):
  * View structs: MatchViewState (alias Phase1ViewState), BodyView,
    ScorePanelView, MatchClockView, GripEdgeView, GripNodeFlash,
    ArrowView, MiniMapView, TextBurst, RefereeFlash.
  * capture_view(match, tick, events, prev_view=None) — pure read.
  * Helpers: damage_band, tint_toward_red, target_to_node_id,
    topdown_hand_positions_mat, owned_hands_by_grasper,
    position_bucket, scaled_duration.
  * Anatomical layout: _BodyPose, _BODY_LAYOUT,
    grip_node_screen_xy, layout_anatomical_pair (+ a
    back-compat _layout_bodies shim for tests).
  * RecordingViewCapture (alias Phase1RecordingRenderer) test fake.
  * TextBurstQueue display-timing helper.

DELETED (src/phase1_viewer.py entirely, ~3000 lines):
  * Phase1AnatomicalRenderer pygame class.
  * WINDOW_W / COL_* / pane-layout constants for the dead viewer.
  * --viewer-phase1 CLI flag from main.py.

Tests retarget to `from viewer_capture import ...`. The HAJ-189
intent/actual force fields on Match (_intent_force,
_actual_force) stay — they're useful capture data even without
the dead renderer.

116/116 viewer tests pass against the new module. main viewer
(`--viewer`) launches cleanly through main.py --help.
